Move SignatureContext to a separate file
This commit addresses initial Simon's PR feedback by moving
the SignatureContext class into a new file SignatureContext.cs
and passing it to the ReadyToRunCodegenNodeFactory instead of
passing the input file name - while we have yet to figure out
the exact design details for R2R module merging, this simple
cleanup can be addressed easily to avoid silently spreading
the input MSIL file name information across the toolchain,
complicating the future merging work.

Thanks

Tomas
